requirements.txt
PyQt5>=5.15.9
pyserial>=3.5
keyboard>=0.13.5
pywin32>=306
numpy>=1.24.3
pyqt5-tools>=5.15.9
python-dotenv>=1.0.0 